Refactor Pragmatic freespin creation logic
- Simplified `PragmaticHttpClient` payload generation by replacing inline DTO construction with unified `gameList` field.
- Implemented `CreateFreespin` in `PragmaticFreespinAdapter`, including preset validation, timestamp conversion, and currency handling.
- Replaced `BigInteger` with `BigDecimal` for `PragmaticBetDto` amount field for better precision.
- Updated `CreateFreespinDto` to include `gameList` and removed deprecated fields.

+++ src/main/kotlin/infrastructure/aggregator/pragmatic/handler/dto/PragmaticBetDto.kt
@@ -1,6 +1,6 @@
 package infrastructure.aggregator.pragmatic.handler.dto
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class PragmaticBetDto(
     val gameSymbol: String,
@@ -11,5 +11,5 @@ data class PragmaticBetDto(
 
     val freeSpinId: String?,
 
-    val amount: BigInteger
+    val amount: BigDecimal
 )
